+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### New features
 
+- Added `\Wizapalce\SDK\Catalog\Review\ReviewService::canUserReviewProduct`
+
 ### Corrections
 
 ## 1.34.0
@@ -17,7 +19,6 @@
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::removeUserToGroup`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::getOrganisationBaskets`
 - Added `\Wizaplace\SDK\Organisation\OrganisationService::addNewUser`
-- Added `\Wizapalce\SDK\Catalog\Review\ReviewService::canUserReviewProduct`
 
 ## 1.33.0